+++ src/main/scala/com/twitter/scalding/JobTest.scala
@@ -4,6 +4,7 @@ import scala.collection.mutable.{Buffer, ListBuffer}
 import scala.annotation.tailrec
 
 import cascading.tuple.Tuple
+import cascading.tuple.TupleEntry
 
 import org.apache.hadoop.mapred.JobConf
 
@@ -21,8 +22,11 @@ object JobTest {
 class JobTest(jobName : String) extends TupleConversions {
   private var argsMap = Map[String, List[String]]()
   private val callbacks = Buffer[() => Unit]()
+  // TODO: Switch the following maps and sets from Source to String keys
+  // to guard for scala equality bugs
   private var sourceMap = Map[Source, Buffer[Tuple]]()
   private var sinkSet = Set[Source]()
+  private var fileSet = Set[String]()
 
   def arg(inArg : String, value : List[String]) = {
     argsMap += inArg -> value
@@ -44,45 +48,70 @@ class JobTest(jobName : String) extends TupleConversions {
     val buffer = new ListBuffer[Tuple]
     sourceMap += s -> buffer
     sinkSet += s
-    callbacks += (() => op(buffer.map{conv(_)}))
+    callbacks += (() => op(buffer.map { tup => conv(new TupleEntry(tup)) }))
     this
   }
 
+  // Simulates the existance of a file so that mode.fileExists returns true.  We
+  // do not simulate the file contents; that should be done through mock
+  // sources.
+  def registerFile(filename : String) = {
+    fileSet += filename
+    this
+  }
+
+
   def run = {
-    Mode.mode = Test(sourceMap)
-    runAll(Job(jobName, new Args(argsMap)))
+    runJob(initJob(false), true)
+    this
+  }
+
+  def runWithoutNext(useHadoop : Boolean = false) = {
+    runJob(initJob(useHadoop), false)
     this
   }
 
   def runHadoop = {
-    Mode.mode = HadoopTest(new JobConf(), sourceMap)
-    runAll(Job(jobName, new Args(argsMap)), true)
+    runJob(initJob(true), true)
     this
   }
 
   // This SITS is unfortunately needed to get around Specs
   def finish : Unit = { () }
 
+  // Registers test files, initializes the global mode, and creates a job.
+  private def initJob(useHadoop : Boolean) : Job = {
+    // Create a global mode to use for testing.
+    val testMode : TestMode =
+      if (useHadoop) {
+        val conf = new JobConf
+        // Set the polling to a lower value to speed up tests:
+        conf.set("jobclient.completion.poll.interval", "100")
+        HadoopTest(conf, sourceMap)
+      } else {
+        Test(sourceMap)
+      }
+    testMode.registerTestFiles(fileSet)
+    Mode.mode = testMode
+
+    // Construct a job.
+    Job(jobName, new Args(argsMap))
+  }
+
   @tailrec
-  final def runAll(job : Job, useHadoop : Boolean = false) : Unit = {
+  private final def runJob(job : Job, runNext : Boolean) : Unit = {
     job.buildFlow.complete
-    job.next match {
-      case Some(nextjob) => runAll(nextjob, useHadoop)
+    val next : Option[Job] = if (runNext) { job.next } else { None }
+    next match {
+      case Some(nextjob) => runJob(nextjob, runNext)
       case None => {
-        if(useHadoop) {
-          sinkSet.foreach{ _.finalizeHadoopTestOutput(Mode.mode) }
+        Mode.mode match {
+          case HadoopTest(_,_) => sinkSet.foreach{ _.finalizeHadoopTestOutput(Mode.mode) }
+          case _ => ()
         }
-        //Now it is time to check the test conditions:
+        // Now it is time to check the test conditions:
         callbacks.foreach { cb => cb() }
       }
     }
   }
-
-  def runWithoutNext = {
-    Mode.mode = Test(sourceMap)
-    Job(jobName, new Args(argsMap)).buildFlow.complete
-    callbacks.foreach { cb => cb() }
-    this
-  }
-
 }