+++ src/main/scala/com/twitter/scalding/RichPipe.scala
@@ -40,9 +40,24 @@ object RichPipe extends FieldConversions with TupleConversions with java.io.Seri
   }
 
   def assignName(p : Pipe) = new Pipe(getNextName, p)
+
+  private val REDUCER_KEY = "mapred.reduce.tasks"
+  /**
+   * Gets the underlying config for this pipe and sets the number of reducers
+   * useful for cascading GroupBy/CoGroup pipes.
+   */
+  def setReducers(p : Pipe, reducers : Int) : Pipe = {
+    if(reducers > 0) {
+      p.getProcessConfigDef()
+        .setProperty(REDUCER_KEY, reducers.toString)
+    } else if(reducers != -1) {
+      throw new IllegalArgumentException("Number of reducers must be non-negative")
+    }
+    p
+  }
 }
 
-class RichPipe(val pipe : Pipe) extends java.io.Serializable {
+class RichPipe(val pipe : Pipe) extends java.io.Serializable with JoinAlgorithms {
   import RichPipe._
 
   // Rename the current pipe
@@ -53,25 +68,6 @@ class RichPipe(val pipe : Pipe) extends java.io.Serializable {
   //them to a fields object
   def project(fields : Fields) = {
     new Each(pipe, fields, new Identity(fields))
-  }
-
-  /*
-   * WARNING! doing a cross product with even a moderate sized pipe can
-   * create ENORMOUS output.  The use-case here is attaching a constant (e.g.
-   * a number or a dictionary or set) to each row in another pipe.
-   * A common use-case comes from a groupAll and reduction to one row,
-   * then you want to send the results back out to every element in a pipe
-   *
-   * This uses joinWithTiny, so tiny pipe is replicated to all Mappers.  If it
-   * is large, this will blow up.  Get it: be foolish here and LOSE IT ALL!
-   *
-   * Use at your own risk.
-   */
-  def crossWithTiny(tiny : Pipe) = {
-    val tinyJoin = tiny.map(() -> '__joinTiny__) { (u:Unit) => 1 }
-    map(() -> '__joinBig__) { (u:Unit) => 1 }
-      .joinWithTiny('__joinBig__ -> '__joinTiny__, tinyJoin)
-      .discard('__joinBig__, '__joinTiny__)
   }
 
   //Discard the given fields, and keep the rest
@@ -92,9 +88,6 @@ class RichPipe(val pipe : Pipe) extends java.io.Serializable {
   def groupBy(f : Fields)(builder : GroupBuilder => GroupBuilder) : Pipe = {
     builder(new GroupBuilder(f)).schedule(pipe.getName, pipe)
   }
-
-  @deprecated("Use groupBy for more consistency with scala.collections API")
-  def group(f : Fields)(builder : GroupBuilder => GroupBuilder) : Pipe = groupBy(f)(builder)
 
   // Returns the set of unique tuples containing the specified fields
   def unique(f : Fields) : Pipe = groupBy(f) { _.size('__uniquecount__) }.project(f)
@@ -140,9 +133,10 @@ class RichPipe(val pipe : Pipe) extends java.io.Serializable {
     new Each(pipe, fromFields, new Identity( toFields ), Fields.SWAP)
   }
 
-  def filter[A:TupleConverter](f : Fields)(fn : (A) => Boolean) : Pipe = {
-    implicitly[TupleConverter[A]].assertArityMatches(f)
-    new Each(pipe, f, new FilterFunction(convertMapFn[A,Boolean](fn)))
+  def filter[A](f : Fields)(fn : (A) => Boolean)
+      (implicit conv : TupleConverter[A]) : Pipe = {
+    conv.assertArityMatches(f)
+    new Each(pipe, f, new FilterFunction(fn, conv))
   }
 
   // If you use a map function that does not accept TupleEntry args,
@@ -170,29 +164,54 @@ class RichPipe(val pipe : Pipe) extends java.io.Serializable {
                 (implicit conv : TupleConverter[A], setter : TupleSetter[T]) : Pipe = {
       conv.assertArityMatches(fs._1)
       setter.assertArityMatches(fs._2)
-      val mf = new MapFunction[T](convertMapFn(fn), fs._2)(setter)
+      val mf = new MapFunction[A,T](fn, fs._2, conv, setter)
       new Each(pipe, fs._1, mf, defaultMode(fs._1, fs._2))
   }
   def mapTo[A,T](fs : (Fields,Fields))(fn : A => T)
                 (implicit conv : TupleConverter[A], setter : TupleSetter[T]) : Pipe = {
       conv.assertArityMatches(fs._1)
       setter.assertArityMatches(fs._2)
-      val mf = new MapFunction[T](convertMapFn(fn), fs._2)(setter)
+      val mf = new MapFunction[A,T](fn, fs._2, conv, setter)
       new Each(pipe, fs._1, mf, Fields.RESULTS)
   }
   def flatMap[A,T](fs : (Fields,Fields))(fn : A => Iterable[T])
                 (implicit conv : TupleConverter[A], setter : TupleSetter[T]) : Pipe = {
       conv.assertArityMatches(fs._1)
       setter.assertArityMatches(fs._2)
-      val mf = new FlatMapFunction[T](convertMapFn(fn), fs._2)(setter)
+      val mf = new FlatMapFunction[A,T](fn, fs._2, conv, setter)
       new Each(pipe, fs._1, mf, defaultMode(fs._1,fs._2))
   }
   def flatMapTo[A,T](fs : (Fields,Fields))(fn : A => Iterable[T])
                 (implicit conv : TupleConverter[A], setter : TupleSetter[T]) : Pipe = {
       conv.assertArityMatches(fs._1)
       setter.assertArityMatches(fs._2)
-      val mf = new FlatMapFunction[T](convertMapFn(fn), fs._2)(setter)
+      val mf = new FlatMapFunction[A,T](fn, fs._2, conv, setter)
       new Each(pipe, fs._1, mf, Fields.RESULTS)
+  }
+
+  /**
+   * This is an analog of the SQL/Excel unpivot function which converts columns of data
+   * into rows of data.  Only the columns given as input fields are expanded in this way.
+   * For this operation to be reversible, you need to keep some unique key on each row.
+   * See GroupBuilder.pivot to reverse this operation assuming you leave behind a grouping key
+   * Example:
+   * pipe.unpivot(('w,'x,'y,'z) -> ('feature, 'value))
+   * takes rows like:
+   * key, w, x, y, z
+   * 1, 2, 3, 4, 5
+   * 2, 8, 7, 6, 5
+   * to:
+   * key, feature, value
+   * 1, w, 2
+   * 1, x, 3
+   * 1, y, 4
+   * etc...
+   */
+  def unpivot(fieldDef : (Fields,Fields)) : Pipe = {
+    assert(fieldDef._2.size == 2, "Must specify exactly two Field names for the results")
+    // toKeyValueList comes from TupleConversions
+    pipe.flatMap(fieldDef)(toKeyValueList)
+      .discard(fieldDef._1)
   }
 
   // Keep at most n elements.  This is implemented by keeping
@@ -201,67 +220,6 @@ class RichPipe(val pipe : Pipe) extends java.io.Serializable {
   def limit(n : Long) = new Each(pipe, new Limit(n))
 
   def debug = new Each(pipe, new Debug())
-
-  @deprecated("Equivalent to joinWithSmaller. Be explicit.")
-  def join(fs :(Fields,Fields), that : Pipe, joiner : Joiner = new InnerJoin) = {
-    joinWithSmaller(fs, that, joiner)
-  }
-
-  private val REDUCER_KEY = "mapred.reduce.tasks"
-
-  /**
-  * Avoid going crazy adding more explicit join modes.  Instead do for some other join
-  * mode with a larger pipe:
-  * .then { pipe => other.
-  *           joinWithSmaller(('other1, 'other2)->('this1, 'this2), pipe, new FancyJoin)
-  *       }
-  */
-  def joinWithSmaller(fs :(Fields,Fields), that : Pipe, joiner : Joiner = new InnerJoin, reducers : Int = -1) = {
-    //Rename these pipes to avoid cascading name conflicts
-    val p = new CoGroup(assignName(pipe), fs._1, assignName(that), fs._2, joiner)
-    if(reducers > 0) {
-      p.getProcessConfigDef()
-        .setProperty(REDUCER_KEY, reducers.toString)
-    } else if(reducers != -1) {
-      throw new IllegalArgumentException("Number of reducers must be non-negative")
-    }
-    p
-  }
-
-  def joinWithLarger(fs : (Fields, Fields), that : Pipe, joiner : Joiner = new InnerJoin, reducers : Int = -1) = {
-    that.joinWithSmaller((fs._2, fs._1), this.pipe, joiner, reducers)
-  }
-
-  def leftJoinWithSmaller(fs :(Fields,Fields), that : Pipe, reducers : Int = -1) = {
-    joinWithSmaller(fs, that, new LeftJoin, reducers)
-  }
-
-  def leftJoinWithLarger(fs :(Fields,Fields), that : Pipe, reducers : Int = -1) = {
-    //We swap the order, and turn left into right:
-    that.joinWithSmaller((fs._2, fs._1), this.pipe, new RightJoin, reducers)
-  }
-
-  @deprecated("Equivalent to leftJoinWithSmaller. Be explicit.")
-  def leftJoin(field_def :(Fields,Fields), that : Pipe) = join(field_def, that, new LeftJoin)
-
-  @deprecated("Equivalent to joinWithSmaller. Be explicit.")
-  def outerJoin(field_def :(Fields,Fields), that : Pipe) = join(field_def, that, new OuterJoin)
-
-  /**
-   * This does an assymmetric join, using cascading's "Join".  This only runs through
-   * this pipe once, and keeps the right hand side pipe in memory (but is spillable).
-   * WARNING: this does not work with outer joins, or right joins (according to
-   * cascading documentation), only inner and left join versions are given.
-   */
-  def joinWithTiny(fs :(Fields,Fields), that : Pipe) = {
-    //Rename these pipes to avoid cascading name conflicts
-    new Join(assignName(pipe), fs._1, assignName(that), fs._2, new InnerJoin)
-  }
-
-  def leftJoinWithTiny(fs :(Fields,Fields), that : Pipe) = {
-    //Rename these pipes to avoid cascading name conflicts
-    new Join(assignName(pipe), fs._1, assignName(that), fs._2, new LeftJoin)
-  }
 
   def write(outsource : Source)(implicit flowDef : FlowDef, mode : Mode) = {
     outsource.write(pipe)(flowDef, mode)
@@ -274,5 +232,33 @@ class RichPipe(val pipe : Pipe) extends java.io.Serializable {
     .map((f, 'total_for_normalize) -> f) { args : (Double, Double) =>
       args._1 / args._2
     }
+  }
+
+  /** Maps the input fields into an output field of type T. For example:
+    *
+    *   pipe.pack[(Int, Int)] (('field1, 'field2) -> 'field3)
+    *
+    * will pack fields 'field1 and 'field2 to field 'field3, as long as 'field1 and 'field2
+    * can be cast into integers. The output field 'field3 will be of tupe (Int, Int)
+    *
+    */
+  def pack[T](fs : (Fields, Fields))(implicit packer : TuplePacker[T]) : Pipe = {
+    val (fromFields, toFields) = fs
+    assert(toFields.size == 1, "Can only output 1 field in pack")
+    pipe.map[TupleEntry, T](fs) { packer.newInstance(_) }
+  }
+
+  /** The opposite of pack. Unpacks the input field of type T into
+    * the output fields. For example:
+    *
+    *   pipe.unpack[(Int, Int)] ('field1 -> ('field2, 'field3))
+    *
+    * will unpack 'field1 into 'field2 and 'field3
+    */
+  def unpack[T](fs : (Fields, Fields))(implicit unpacker : TupleUnpacker[T]) : Pipe = {
+    val (fromFields, toFields) = fs
+    assert(fromFields.size == 1, "Can only take 1 input field in unpack")
+    val setter = unpacker.newSetter(toFields)
+    pipe.map[T, Tuple](fs) { input : T => setter(input) }
   }
 }