+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Scalding #
 
+### Version 0.4.0 ###
+* ISSUE 38: Fix bug with hash code collisions of Source objects
+* ISSUE 36: Some cleanups of reduce and Operations
+* ISSUE 35: Split RichPipe join methods into their own trait
+* ISSUE 34: Adds Pivot/Unpivot
+* ISSUE 33: Add pack and unpack methods to RichPipe
+* ISSUE 32: Refactors reducer setting into RichPipe
+* ISSUE 31: Implemented Mode.fileExists
+* ISSUE 28: Simplifies TupleConverter
+
 ### Version 0.3.5 ###
 * ISSUE 21: move JobTest into main
 * ISSUE 20: Adding a source for the most recent good date path.